+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Select and move holdings with items or items to another instance. Refs UIIN-1098.
 * Move items between holdings. Refs UIIN-1097.
+* Disable fields in Preceding/Succeeding section, when Instance is controlled by Source MARC. Refs UIIN-1128.
 
 ## [3.0.1] (IN PROGRESS)